[GAME] Ajout du timer

+++ app/src/main/java/com/example/towerofhanoi/model/GameListener.java
@@ -5,6 +5,6 @@ public interface GameListener {
 
     void onMovementsChanged(int movementCount);
 
-    void onTimeChanged(int time);
+    void onTimeChanged(long time);
 
 }
